Task & User Management
- Add link between user & task in controller
- Add authorization to edit & delete to User and Task
- Add homepage link to title website

+++ src/AppBundle/Controller/TaskController.php
@@ -23,6 +23,9 @@ class TaskController extends Controller
      */
     public function createAction(Request $request)
     {
+        if (!$this->get('security.authorization_checker')->isGranted('IS_AUTHENTICATED_FULLY')) {
+            throw $this->createAccessDeniedException();
+        }
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
 
@@ -30,6 +33,9 @@ class TaskController extends Controller
 
         if ($form->isValid()) {
             $em = $this->getDoctrine()->getManager();
+
+            // ******** Add User link with a task ******** //
+            $task->setUser($this->get('security.token_storage')->getToken()->getUser());
 
             $em->persist($task);
             $em->flush();
@@ -47,22 +53,28 @@ class TaskController extends Controller
      */
     public function editAction(Task $task, Request $request)
     {
-        $form = $this->createForm(TaskType::class, $task);
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+        if($user == $task->getUser() or $user->getRoles()[0] == "ROLE_ADMIN"){
+            $form = $this->createForm(TaskType::class, $task);
 
-        $form->handleRequest($request);
+            $form->handleRequest($request);
 
-        if ($form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            $this->addFlash('success', 'La tâche a bien été modifiée.');
+                $this->addFlash('success', 'La tâche a bien été modifiée.');
 
-            return $this->redirectToRoute('task_list');
+                return $this->redirectToRoute('task_list');
+            }
+
+            return $this->render('task/edit.html.twig', [
+                'form' => $form->createView(),
+                'task' => $task,
+            ]);
         }
+        $this->addFlash('error', "Vous n'êtes pas autorisé à accèder à cette page");
 
-        return $this->render('task/edit.html.twig', [
-            'form' => $form->createView(),
-            'task' => $task,
-        ]);
+        return $this->redirectToRoute('task_list');
     }
 
     /**
@@ -83,11 +95,17 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        $user = $this->get('security.token_storage')->getToken()->getUser();
+        if($user == $task->getUser() or $user->getRoles()[0] == "ROLE_ADMIN") {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+
+            return $this->redirectToRoute('task_list');
+        }
+        $this->addFlash('error', "Vous n'êtes pas autorisé à effectuer cette action");
 
         return $this->redirectToRoute('task_list');
     }
